update member nav; move navItem to shared
